+++ app/lib/pages/home_page.dart
@@ -10,6 +10,8 @@ import 'package:flutter_quill/widgets/controller.dart';
 import 'package:flutter_quill/widgets/default_styles.dart';
 import 'package:flutter_quill/widgets/editor.dart';
 import 'package:flutter_quill/widgets/toolbar.dart';
+import 'package:path/path.dart';
+import 'package:path_provider/path_provider.dart';
 import 'package:tuple/tuple.dart';
 
 import 'read_only_page.dart';
@@ -103,41 +105,44 @@ class _HomePageState extends State<HomePage> {
               focusNode: _focusNode,
               autoFocus: false,
               readOnly: false,
+              placeholder: 'Add content',
               enableInteractiveSelection: true,
               expands: false,
               padding: EdgeInsets.zero,
               customStyles: DefaultStyles(
-                  h1: DefaultTextBlockStyle(
-                      TextStyle(
-                        fontSize: 32.0,
-                        color: Colors.black,
-                        height: 1.15,
-                        fontWeight: FontWeight.w300,
-                      ),
-                      Tuple2(16.0, 0.0),
-                      Tuple2(0.0, 0.0),
-                      null),
-                  sizeSmall: TextStyle(fontSize: 9.0)),
+                h1: DefaultTextBlockStyle(
+                    TextStyle(
+                      fontSize: 32.0,
+                      color: Colors.black,
+                      height: 1.15,
+                      fontWeight: FontWeight.w300,
+                    ),
+                    Tuple2(16.0, 0.0),
+                    Tuple2(0.0, 0.0),
+                    null),
+                sizeSmall: TextStyle(fontSize: 9.0),
+              ),
             ),
           ),
           Container(
-            padding:
-                EdgeInsets.only(top: MediaQuery.of(context).size.height * 0.9),
             child: QuillToolbar.basic(
                 controller: _controller,
-                uploadFileCallback: _fakeUploadImageCallBack),
-          )
+                onImagePickCallback: _onImagePickCallback),
+          ),
         ],
       ),
     );
   }
 
-  Future<String> _fakeUploadImageCallBack(File file) async {
-    print(file);
-    var completer = new Completer<String>();
-    completer.complete(
-        'https://user-images.githubusercontent.com/122956/72955931-ccc07900-3d52-11ea-89b1-d468a6e2aa2b.png');
-    return completer.future;
+  // Renders the image picked by imagePicker from local file storage
+  // You can also upload the picked image to any server (eg : AWS s3 or Firebase) and then return the uploaded image URL
+  Future<String> _onImagePickCallback(File file) async {
+    if (file == null) return null;
+    // Copies the picked file from temporary cache to applications directory
+    Directory appDocDir = await getApplicationDocumentsDirectory();
+    File copiedFile =
+        await file.copy('${appDocDir.path}/${basename(file.path)}');
+    return copiedFile.path.toString();
   }
 
   Widget _buildMenuBar(BuildContext context) {
@@ -156,7 +161,7 @@ class _HomePageState extends State<HomePage> {
 
   void _readOnly() {
     Navigator.push(
-      context,
+      super.context,
       MaterialPageRoute(
         builder: (BuildContext context) => ReadOnlyPage(),
       ),