+++ lib/models/documents/document.dart
@@ -237,6 +237,22 @@ class Document {
       _root.remove(node);
     }
   }
+
+  bool isEmpty() {
+    if (root.children.length != 1) {
+      return false;
+    }
+
+    final Node node = root.children.first;
+    if (!node.isLast) {
+      return false;
+    }
+
+    Delta delta = node.toDelta();
+    return delta.length == 1 &&
+        delta.first.data == '\n' &&
+        delta.first.key == 'insert';
+  }
 }
 
 enum ChangeSource {