+++ lib/widgets/raw_editor.dart
@@ -1,3 +1,5 @@
+import 'dart:convert';
+
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/gestures.dart';
@@ -32,6 +34,7 @@ class RawEditor extends StatefulWidget {
   final bool scrollable;
   final EdgeInsetsGeometry padding;
   final bool readOnly;
+  final String placeholder;
   final ValueChanged<String> onLaunchUrl;
   final ToolbarOptions toolbarOptions;
   final bool showSelectionHandles;
@@ -58,6 +61,7 @@ class RawEditor extends StatefulWidget {
       this.scrollable,
       this.padding,
       this.readOnly,
+      this.placeholder,
       this.onLaunchUrl,
       this.toolbarOptions,
       this.showSelectionHandles,
@@ -374,7 +378,7 @@ class RawEditorState extends EditorState
           inputType: TextInputType.multiline,
           readOnly: widget.readOnly,
           obscureText: false,
-          autocorrect: false,
+          autocorrect: true,
           inputAction: TextInputAction.newline,
           keyboardAppearance: widget.keyboardAppearance,
           textCapitalization: widget.textCapitalization,
@@ -503,13 +507,21 @@ class RawEditorState extends EditorState
     _focusAttachment.reparent();
     super.build(context);
 
+    Document _doc = widget.controller.document;
+    if (_doc.isEmpty() &&
+        !widget.focusNode.hasFocus &&
+        widget.placeholder != null) {
+      _doc = Document.fromJson(jsonDecode(
+          '[{"attributes":{"placeholder":true},"insert":"${widget.placeholder}\\n"}]'));
+    }
+
     Widget child = CompositedTransformTarget(
       link: _toolbarLayerLink,
       child: Semantics(
         child: _Editor(
           key: _editorKey,
-          children: _buildChildren(context),
-          document: widget.controller.document,
+          children: _buildChildren(_doc, context),
+          document: _doc,
           selection: widget.controller.selection,
           hasFocus: _hasFocus,
           textDirection: _textDirection,
@@ -562,30 +574,13 @@ class RawEditorState extends EditorState
     requestKeyboard();
   }
 
-  _buildChildren(BuildContext context) {
+  _buildChildren(Document doc, BuildContext context) {
     final result = <Widget>[];
     Map<int, int> indentLevelCounts = {};
-    for (Node node in widget.controller.document.root.children) {
+    for (Node node in doc.root.children) {
       if (node is Line) {
-        TextLine textLine = TextLine(
-          line: node,
-          textDirection: _textDirection,
-          embedBuilder: widget.embedBuilder,
-          styles: _styles,
-        );
-        EditableTextLine editableTextLine = EditableTextLine(
-            node,
-            null,
-            textLine,
-            0,
-            _getVerticalSpacingForLine(node, _styles),
-            _textDirection,
-            widget.controller.selection,
-            widget.selectionColor,
-            widget.enableInteractiveSelection,
-            _hasFocus,
-            MediaQuery.of(context).devicePixelRatio,
-            _cursorCont);
+        EditableTextLine editableTextLine =
+            _getEditableTextLineFromNode(node, context);
         result.add(editableTextLine);
       } else if (node is Block) {
         Map<String, Attribute> attrs = node.style.attributes;
@@ -610,6 +605,30 @@ class RawEditorState extends EditorState
       }
     }
     return result;
+  }
+
+  EditableTextLine _getEditableTextLineFromNode(
+      Line node, BuildContext context) {
+    TextLine textLine = TextLine(
+      line: node,
+      textDirection: _textDirection,
+      embedBuilder: widget.embedBuilder,
+      styles: _styles,
+    );
+    EditableTextLine editableTextLine = EditableTextLine(
+        node,
+        null,
+        textLine,
+        0,
+        _getVerticalSpacingForLine(node, _styles),
+        _textDirection,
+        widget.controller.selection,
+        widget.selectionColor,
+        widget.enableInteractiveSelection,
+        _hasFocus,
+        MediaQuery.of(context).devicePixelRatio,
+        _cursorCont);
+    return editableTextLine;
   }
 
   Tuple2<double, double> _getVerticalSpacingForLine(